errot修正中

+++ project/Commons/Object3dCommon.cpp
@@ -295,7 +295,7 @@ void Object3dCommon::MakeSkeltonPSO()
 	inputElementDescs[4].AlignedByteOffset = D3D12_APPEND_ALIGNED_ELEMENT;
 
 	inputLayoutDesc.pInputElementDescs = inputElementDescs.data();
-	inputLayoutDesc.NumElements = inputElementDescs.size();
+	inputLayoutDesc.NumElements = UINT(inputElementDescs.size());
 
 	D3D12_BLEND_DESC blendDesc{};
 	blendDesc.RenderTarget[0].RenderTargetWriteMask = D3D12_COLOR_WRITE_ENABLE_ALL;
